+++ java/src/main/java/io/cucumber/cucumberexpressions/GroupBuilder.java
@@ -20,6 +20,7 @@ final class GroupBuilder {
         this.startIndex = startIndex;
     }
 
+
     void add(GroupBuilder groupBuilder) {
         groupBuilders.add(groupBuilder);
     }
@@ -30,7 +31,25 @@ final class GroupBuilder {
         for (GroupBuilder childGroupBuilder : groupBuilders) {
             children.add(childGroupBuilder.build(matcher, groupIndices));
         }
-        return new Group(matcher.group(groupIndex), matcher.start(groupIndex), matcher.end(groupIndex), children);
+        return new Group(
+                matcher.group(groupIndex), //
+                matcher.start(groupIndex),  //
+                matcher.end(groupIndex),  //
+                children.isEmpty() ? null : children //
+        );
+    }
+
+    List<Group> toGroups() {
+        List<Group> list = new ArrayList<>();
+        for (GroupBuilder child : groupBuilders) {
+            List<Group> groups = child.toGroups();
+            list.add(new Group(
+                    child.getSource(),
+                    child.getStartIndex(),
+                    child.getEndIndex(),
+                    groups.isEmpty() ? null : groups));
+        }
+        return list;
     }
 
     void setNonCapturing() {
@@ -52,7 +71,7 @@ final class GroupBuilder {
     }
 
     String getSource() {
-        return requireNonNull(source); 
+        return requireNonNull(source);
     }
 
     void setSource(String source) {