+++ java/src/test/java/io/cucumber/cucumberexpressions/TreeRegexpTest.java
@@ -10,6 +10,7 @@ import static java.util.Arrays.asList;
 
 import static java.util.Objects.requireNonNull;
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertNull;
 
 public class TreeRegexpTest {
@@ -30,10 +31,10 @@ public class TreeRegexpTest {
         TreeRegexp tr = new TreeRegexp("(a(b(c))(d))");
         Group g = tr.match("abcd");
         requireNonNull(g);
-        assertEquals("abcd", g.getChildren().get(0).getValue());
-        assertEquals("bc", g.getChildren().get(0).getChildren().get(0).getValue());
-        assertEquals("c", g.getChildren().get(0).getChildren().get(0).getChildren().get(0).getValue());
-        assertEquals("d", g.getChildren().get(0).getChildren().get(1).getValue());
+        assertEquals("abcd", g.getChildren().get().get(0).getValue());
+        assertEquals("bc", g.getChildren().get().get(0).getChildren().get().get(0).getValue());
+        assertEquals("c", g.getChildren().get().get(0).getChildren().get().get(0).getChildren().get().get(0).getValue());
+        assertEquals("d", g.getChildren().get().get(0).getChildren().get().get(1).getValue());
     }
 
     @Test
@@ -42,7 +43,7 @@ public class TreeRegexpTest {
         Group g = tr.match("abc");
         requireNonNull(g);
         assertEquals("abc", g.getValue());
-        assertEquals(1, g.getChildren().size());
+        assertEquals(1, g.getChildren().get().size());
     }
 
     @Test
@@ -51,7 +52,7 @@ public class TreeRegexpTest {
         Group g = tr.match("aBc");
         requireNonNull(g);
         assertEquals("aBc", g.getValue());
-        assertEquals(1, g.getChildren().size());
+        assertEquals(1, g.getChildren().get().size());
     }
 
     @Test
@@ -60,8 +61,8 @@ public class TreeRegexpTest {
         Group g = tr.match("abc");
         requireNonNull(g);
         assertEquals("abc", g.getValue());
-        assertEquals(1, g.getChildren().size());
-        assertEquals("bc", g.getChildren().get(0).getValue());
+        assertEquals(1, g.getChildren().get().size());
+        assertEquals("bc", g.getChildren().get().get(0).getValue());
     }
 
     @Test
@@ -70,8 +71,8 @@ public class TreeRegexpTest {
         Group g = tr.match("abc");
         requireNonNull(g);
         assertEquals("abc", g.getValue());
-        assertEquals(1, g.getChildren().size());
-        assertEquals("bc", g.getChildren().get(0).getValue());
+        assertEquals(1, g.getChildren().get().size());
+        assertEquals("bc", g.getChildren().get().get(0).getValue());
     }
 
     @Test
@@ -80,8 +81,8 @@ public class TreeRegexpTest {
         Group g = tr.match("abc");
         requireNonNull(g);
         assertEquals("abc", g.getValue());
-        assertEquals(1, g.getChildren().size());
-        assertEquals("bc", g.getChildren().get(0).getValue());
+        assertEquals(1, g.getChildren().get().size());
+        assertEquals("bc", g.getChildren().get().get(0).getValue());
     }
 
     @Test
@@ -90,8 +91,8 @@ public class TreeRegexpTest {
         Group g = tr.match("abc");
         requireNonNull(g);
         assertEquals("abc", g.getValue());
-        assertEquals(1, g.getChildren().size());
-        assertEquals("c", g.getChildren().get(0).getValue());
+        assertEquals(1, g.getChildren().get().size());
+        assertEquals("c", g.getChildren().get().get(0).getValue());
     }
 
     @Test
@@ -100,8 +101,8 @@ public class TreeRegexpTest {
         Group g = tr.match("abc");
         requireNonNull(g);
         assertEquals("abc", g.getValue());
-        assertEquals(1, g.getChildren().size());
-        assertEquals("b", g.getChildren().get(0).getValue());
+        assertEquals(1, g.getChildren().get().size());
+        assertEquals("b", g.getChildren().get().get(0).getValue());
     }
 
     @Test
@@ -109,7 +110,7 @@ public class TreeRegexpTest {
         TreeRegexp tr = new TreeRegexp("^Something( with an optional argument)?");
         Group g = tr.match("Something");
         requireNonNull(g);
-        assertNull(g.getChildren().get(0).getValue());
+        assertNull(g.getChildren().get().get(0).getValue());
     }
 
     @Test
@@ -117,15 +118,15 @@ public class TreeRegexpTest {
         TreeRegexp tr = new TreeRegexp("^A (\\d+) thick line from ((\\d+),\\s*(\\d+),\\s*(\\d+)) to ((\\d+),\\s*(\\d+),\\s*(\\d+))");
         Group g = tr.match("A 5 thick line from 10,20,30 to 40,50,60");
         requireNonNull(g);
-        assertEquals("5", g.getChildren().get(0).getValue());
-        assertEquals("10,20,30", g.getChildren().get(1).getValue());
-        assertEquals("10", g.getChildren().get(1).getChildren().get(0).getValue());
-        assertEquals("20", g.getChildren().get(1).getChildren().get(1).getValue());
-        assertEquals("30", g.getChildren().get(1).getChildren().get(2).getValue());
-        assertEquals("40,50,60", g.getChildren().get(2).getValue());
-        assertEquals("40", g.getChildren().get(2).getChildren().get(0).getValue());
-        assertEquals("50", g.getChildren().get(2).getChildren().get(1).getValue());
-        assertEquals("60", g.getChildren().get(2).getChildren().get(2).getValue());
+        assertEquals("5", g.getChildren().get().get(0).getValue());
+        assertEquals("10,20,30", g.getChildren().get().get(1).getValue());
+        assertEquals("10", g.getChildren().get().get(1).getChildren().get().get(0).getValue());
+        assertEquals("20", g.getChildren().get().get(1).getChildren().get().get(1).getValue());
+        assertEquals("30", g.getChildren().get().get(1).getChildren().get().get(2).getValue());
+        assertEquals("40,50,60", g.getChildren().get().get(2).getValue());
+        assertEquals("40", g.getChildren().get().get(2).getChildren().get().get(0).getValue());
+        assertEquals("50", g.getChildren().get().get(2).getChildren().get().get(1).getValue());
+        assertEquals("60", g.getChildren().get().get(2).getChildren().get().get(2).getValue());
     }
 
     @Test
@@ -134,8 +135,8 @@ public class TreeRegexpTest {
         Group g = tr.match("the stdout");
         requireNonNull(g);
         assertEquals("the stdout", g.getValue());
-        assertNull(g.getChildren().get(0).getValue());
-        assertEquals(1, g.getChildren().size());
+        assertNull(g.getChildren().get().get(0).getValue());
+        assertEquals(1, g.getChildren().get().size());
     }
 
     @Test
@@ -143,7 +144,7 @@ public class TreeRegexpTest {
         TreeRegexp tr = new TreeRegexp("(?:a)(:b)(\\?c)(d)");
         Group g = tr.match("a:b?cd");
         requireNonNull(g);
-        assertEquals(3, g.getChildren().size());
+        assertEquals(3, g.getChildren().get().size());
     }
 
     @Test
@@ -151,7 +152,7 @@ public class TreeRegexpTest {
         TreeRegexp tr = new TreeRegexp("foo\\\\(bar|baz)");
         Group g = tr.match("foo\\bar");
         requireNonNull(g);
-        assertEquals(1, g.getChildren().size());
+        assertEquals(1, g.getChildren().get().size());
     }
 
     @Test
@@ -159,7 +160,7 @@ public class TreeRegexpTest {
         TreeRegexp tr = new TreeRegexp("^I go to '/(.+)'$");
         Group g = tr.match("I go to '/hello'");
         requireNonNull(g);
-        assertEquals(1, g.getChildren().size());
+        assertEquals(1, g.getChildren().get().size());
     }
 
     @Test
@@ -167,7 +168,7 @@ public class TreeRegexpTest {
         TreeRegexp tr = new TreeRegexp("^(\\d) (\\w+) (\\w+)$");
         Group g = tr.match("2 you привет");
         requireNonNull(g);
-        assertEquals(3, g.getChildren().size());
+        assertEquals(3, g.getChildren().get().size());
     }
 
     @Test
@@ -175,10 +176,10 @@ public class TreeRegexpTest {
         TreeRegexp tr = new TreeRegexp("^the step \"([^\"]*)\" has status \"([^\"]*)\"$");
         Group g = tr.match("the step \"a pending step\" has status \"pending\"");
         requireNonNull(g);
-        assertEquals(10, g.getChildren().get(0).getStart());
-        assertEquals(24, g.getChildren().get(0).getEnd());
-        assertEquals(38, g.getChildren().get(1).getStart());
-        assertEquals(45, g.getChildren().get(1).getEnd());
+        assertEquals(10, g.getChildren().get().get(0).getStart());
+        assertEquals(24, g.getChildren().get().get(0).getEnd());
+        assertEquals(38, g.getChildren().get().get(1).getStart());
+        assertEquals(45, g.getChildren().get().get(1).getEnd());
     }
 
     @Test
@@ -187,8 +188,8 @@ public class TreeRegexpTest {
         Group g = tr.match("drawings: FU(BAR)");
         requireNonNull(g);
         assertEquals("drawings: FU(BAR)", g.getValue());
-        assertEquals("FU(BAR)", g.getChildren().get(0).getValue());
-        assertEquals(0, g.getChildren().get(0).getChildren().size());
+        assertEquals("FU(BAR)", g.getChildren().get().get(0).getValue());
+        assertFalse(g.getChildren().get().get(0).getChildren().isPresent());
     }
 
     @Test
@@ -205,7 +206,7 @@ public class TreeRegexpTest {
         Group g = tr.match("hello");
         requireNonNull(g);
         assertEquals("hello", g.getValue());
-        assertEquals(0, g.getChildren().size());
+        assertFalse(g.getChildren().isPresent());
     }
 
     @Test
@@ -214,7 +215,7 @@ public class TreeRegexpTest {
         Group g = tr.match("hello");
         requireNonNull(g);
         assertEquals("hello", g.getValue());
-        assertEquals(0, g.getChildren().size());
+        assertFalse(g.getChildren().isPresent());
     }
 
     @Test
@@ -223,7 +224,7 @@ public class TreeRegexpTest {
         Group g = tr.match("");
         requireNonNull(g);
         assertEquals("", g.getValue());
-        assertEquals(1, g.getChildren().size());
+        assertEquals(1, g.getChildren().get().size());
     }
 
     @Test
@@ -232,7 +233,7 @@ public class TreeRegexpTest {
         Group g = tr.match("");
         requireNonNull(g);
         assertEquals("", g.getValue());
-        assertEquals(0, g.getChildren().size());
+        assertFalse(g.getChildren().isPresent());
     }
 
     @Test
@@ -241,7 +242,7 @@ public class TreeRegexpTest {
         Group g = tr.match("");
         requireNonNull(g);
         assertEquals("", g.getValue());
-        assertEquals(0, g.getChildren().size());
+        assertFalse(g.getChildren().isPresent());
     }
 
     @Test